Refactors Serverless Function
What
---

I refactored the serverless function to remove the manual JSON
marshaling, only accept POST requests, and UnMarshal the request body
(JSON) to a concrete type.

Why
---

I want to add a database (supabase) to store all the yells and in
preparation for that I needed to clean up the code a bit.